deleted files *.styled.ts and refactor
