+++ src/components/UserSettings/UserSettings.tsx
@@ -1,14 +1,12 @@
 import React from "react";
-import { Typography, Button, Divider, Box } from "@mui/material";
+import { Typography, Button, Divider, Box, InputLabel, TextField } from "@mui/material";
 import { useDispatch } from "react-redux";
 import { showNotification } from "@/redux/notification/notificationSlice";
-import {
-  StyledFormWrapper,
-  StyledContainer,
-} from "@components/UserSettings/UserSettings.styled";
 import { useDeactivateAccountMutation } from "@/redux/userSettings/userSettingsApiSlice";
 import ChangePasswordForm from "@components/ChangePasswordForm/ChangePasswordForm.tsx";
 import UpdateProfileForm from "@components/UpdateProfileForm/UpdateProfileForm";
+import { styled } from "@mui/material/styles";
+import { alpha } from "@mui/system";
 
 const UserSettings: React.FC = () => {
   const dispatch = useDispatch();
@@ -58,3 +56,39 @@ const UserSettings: React.FC = () => {
 };
 
 export default UserSettings;
+
+export const StyledFormWrapper = styled("div")(({ theme }) => ({
+  display: "flex",
+  justifyContent: "center",
+  padding: theme.spacing(8, 0),
+}));
+
+export const StyledContainer = styled("div")(() => ({
+  width: "600px",
+}));
+
+export const FormRow = styled("div")(({ theme }) => ({
+  display: "flex",
+  margin: "12px 0",
+  gap: theme.spacing(2),
+}));
+
+export const FormInputLabel = styled(InputLabel)(({ theme }) => ({
+  color: `${theme.palette.primary.dark}`,
+  fontSize: "16px",
+  fontStyle: "normal",
+  fontWeight: 400,
+  paddingBottom: "7px",
+}));
+
+export const FormInput = styled(TextField)(({ theme }) => ({
+  "& .MuiInputBase-input": {
+    color: theme.palette.text.secondary,
+    borderRadius: "12px",
+    border: `1px solid ${alpha(theme.palette.primary.light, 0.35)}`,
+  },
+  "& .MuiOutlinedInput-root ": {
+    borderRadius: "12px",
+    border: "none",
+  },
+}));
